remove compile error and warning of motor module

+++ motor/inc/math.h
@@ -6,8 +6,8 @@
 
 #define maxShort (1<<16)
 #define midShort (1<<15)
-#define divSQRT_3 (0.57735026918963*midShort) /* 1/sqrt(3) */
-#define sqrt3DIV_2 (0.86602540378444*midShort)
+#define divSQRT_3 ((short)(0.57735026918963*midShort)) /* 1/sqrt(3) */
+#define sqrt3DIV_2 ((short)(0.86602540378444*midShort))
 
 /*loopup table to get sin(theta), cos(theta) at the same time in order to increase speed*/
 /*theta: unit: 0001 = pi/2/(2^13) = 0.01degree*/
